requirements.txt
annotated-types==0.7.0
asgiref==3.9.2
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Django==5.2.6
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.41.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.10
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.5
pytesseract==0.3.13
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
